re-organise. Significant changes to architecture
