--- ChromProcess/Loading/custom/file_import.py
@@ -1,133 +1,0 @@
-'''
-Functions for importing chromatography data.
-'''
-def read_local_assignments(file):
-    '''
-    Read a comma-separated values file containing assignments of compounds to
-    retention time boundaries.
-
-    File format:
-        compound_1_name,lower bound (float), upper bound (float), 
-        compound_2_name,lower bound (float), upper bound (float), 
-        etc.
-
-    Parameters
-    ----------
-    file: str
-        Path to file
-
-    Returns
-    -------
-    modified_bounds: dict
-        modified_bounds['compound_1_name'] = [lower: float, upper: float]
-    '''
-    import os
-
-    modified_bounds = {}
-
-    if not os.path.exists(file):
-        print(f'''ChromProcess.file_import.read_local_assignments:
-        Local assignments file at {file} not found''')
-        print('''ChromProcess.file_import.read_local_assignments:
-        Returning empty dict''')
-        return modified_bounds
-    else:
-        with open(file, 'r') as f:
-            for line in f:
-                spl = line.strip('\n').split(',')
-                spl = [x for x in spl if x != '']
-                if len(spl) == 0:
-                    pass
-                elif 'set_IS_pos' in line:
-                    modified_bounds[spl[0]] = [float(spl[1])]
-                else:
-                    modified_bounds[spl[0]] = [float(spl[1]), float(spl[2])]
-
-    return modified_bounds
-
-def read_mass_spectra_report(file):
-    '''
-    Read the mass spectra from a file generated by 
-    Classes.Chromatogram.write_peak_mass_spectra()
-
-    Parameters
-    ----------
-    file: str
-        path to file
-
-    Returns
-    --------
-    mass_spectra: list
-        A list of ChromProcess MassSpectrum objects.
-    '''
-
-    import numpy as np
-    from ChromProcess import Classes
-    from ChromProcess import simple_functions as s_f
-
-    mass_spectra = []
-
-    fltln = lambda x: [float(e) for e in x.strip('\n').split(',')[1:]
-                       if e != '']
-
-    mz = np.array([0.0,0.0])
-    r_a = np.array([0.0,0.0])
-    rt = 0.0
-
-    with open(file, 'r') as f:
-
-        make_ms = False
-        for line in f:
-            if 'Peak retention time' in line:
-                read = line.strip('\n').split(',')
-                rt = float(read[1])
-            if 'm/z' in line:
-                if s_f.isfloat(line.split(',')[1]):
-                    read = fltln(line)
-                    mz = np.array(read)
-                else:
-                    mz = np.array([0.0])
-
-            if 'relative abundance' in line:
-                if s_f.isfloat(line.split(',')[1]):
-                    read = fltln(line)
-                    r_a = np.array(read)
-                else:
-                    r_a = np.array([0.0])
-
-                make_ms = True
-
-            if make_ms:
-                mass_spectra.append(
-                    Classes.MassSpectrum(file, mz, r_a, pos = round(rt,3))
-                )
-                make_ms = False
-
-    return mass_spectra
-
-def read_point_removals_file(fname):
-    '''
-    For reading a file containing datasets and point indices to remove in the
-    data sets.
-
-    Parameters
-    ----------
-    fname: str
-        Path to the file
-
-    Returns
-    -------
-    point_removals: dict
-        dictionary of point removal indices. keys are experiment codes,
-        list of ints are items.
-    '''
-    point_removals = {}
-    with open(fname, 'r') as f:
-        for c,line in enumerate(f):
-            if c == 0:
-                pass
-            else:
-                ins = line.strip('\n').split(',')
-                point_removals[ins[0]] = [int(x) for x in ins[1:] if x!='']
-
-    return point_removals
