+++ ChromProcess/Loading/custom/mass_spectra_from_csv.py
@@ -1,0 +1,61 @@
+import numpy as np
+
+from ChromProcess import Classes
+from ChromProcess.Utils import simple_functions as s_f
+
+def mass_spectra_from_csv(file):
+    '''
+    Read the mass spectra from a file generated by 
+    Classes.Chromatogram.write_peak_mass_spectra()
+
+    Parameters
+    ----------
+    file: str
+        path to file
+
+    Returns
+    --------
+    mass_spectra: list
+        A list of ChromProcess MassSpectrum objects.
+    '''
+
+    mass_spectra = []
+
+    fltln = lambda x: [float(e) for e in x.strip('\n').split(',')[1:]
+                       if e != '']
+
+    mz = np.array([0.0,0.0])
+    r_a = np.array([0.0,0.0])
+    rt = 0.0
+
+    with open(file, 'r') as f:
+
+        make_ms = False
+        for line in f:
+            if 'Peak retention time' in line:
+                read = line.strip('\n').split(',')
+                rt = float(read[1])
+            if 'm/z' in line:
+                if s_f.isfloat(line.split(',')[1]):
+                    read = fltln(line)
+                    mz = np.array(read)
+                else:
+                    mz = np.array([0.0])
+
+            if 'relative abundance' in line:
+                if s_f.isfloat(line.split(',')[1]):
+                    read = fltln(line)
+                    r_a = np.array(read)
+                else:
+                    r_a = np.array([0.0])
+
+                make_ms = True
+
+            if make_ms:
+                mass_spectra.append(
+                    Classes.MassSpectrum(file, mz, r_a, pos = round(rt,3))
+                )
+                make_ms = False
+
+    return mass_spectra
+
