+++ ChromProcess/Loading/parsers/parsers.py
@@ -11,7 +11,7 @@ def parse_text_columns(text, point_delimiter, ordinal_delimiter):
     text: str
         Text to be parsed.
     point_delimiter: str
-        Delimiter for separating point coordintes. 
+        Delimiter for separating point coordintes.
     ordinal_delimiter: str
         Delimiter for separating x,y pairs
     Returns
@@ -30,38 +30,3 @@ def parse_text_columns(text, point_delimiter, ordinal_delimiter):
     data = list(map(list, zip(*points)))
 
     return data
-
-def chrom_from_text(x_values, y_values, x_unit, y_unit, filename):
-    '''
-    Create a chromatogram object and insert time and signal
-    information into it.
-
-    Parameters
-    ----------
-    x_values: list
-    y_values: list
-    x_unit: str
-    y_unit: str
-
-    Returns
-    -------
-    chrom: ChromProcess Chromatogram object
-    '''
-    assert isinstance(x_values, list), "x_values arg should be a list"
-    assert isinstance(y_values, list), "y_values arg should be list"
-    assert isinstance(x_unit, str), "x_unit arg should be str"
-    assert isinstance(y_unit, str), "y_unit arg should be str"
-
-    chrom = Classes.Chromatogram()
-    chrom.filename = filename
-
-    chrom.x_unit = x_unit
-    chrom.y_unit = y_unit
-
-    x_values = [float(x) for x in x_values]
-    y_values = [float(y) for y in y_values]
-
-    chrom.time = np.array(x_values)
-    chrom.signal = np.array(y_values)
-
-    return chrom
